Split the monolithic mox_test.py and fix a mox.expect isolation bug
Break the 4,629-line test/mox_test.py into focused per-concern modules
(max ~1,200 lines each):

- test_exceptions.py, test_comparators.py
- test_mock_objects.py / test_mock_object_contextmanager.py
- test_mox.py / test_mox_contextmanager.py
- test_mox_testbase.py
- mox_test_fixtures_helper.py (shared fixture classes; ignored by collection)

The split is a verbatim move - the set of collected test node ids is
byte-for-byte identical to before (379 tests). `TestClass` gets
`__test__ = False`, removing the long-standing PytestCollectionWarning.

Splitting surfaced a latent cross-test isolation bug: `mox.expect` is a
module-level singleton, and `with mox.expect(stub):` left `stubs`/`mox_obj`
set on it, so a later bare `with mox.expect:` skipped the global replay and
mocks were never put into replay mode. It only manifested once the
context-manager tests ran after mox_test_string_imports.py (which the old
file ordering happened to avoid). Expect.__exit__ now resets that transient
state in a finally block, with a regression test.

Full suite: 472 passed; coverage 93%; zero collection warnings.
